Rename package module from onboarding to daskOnboarding
