+++ packages/backend/src/index.ts
@@ -36,7 +36,7 @@ import explore from './plugins/explore';
 import lighthouse from './plugins/lighthouse';
 import kubernetes from './plugins/kubernetes';
 import devTools from './plugins/devtools';
-import onboarding from './plugins/onboarding';
+import daskOnboarding from './plugins/dask-onboarding';
 
 function makeCreateEnv(config: Config) {
   const root = getRootLogger();
@@ -97,7 +97,7 @@ async function main() {
   const lighthouseEnv = useHotMemoize(module, () => createEnv('lighthouse'));
   const kubernetesEnv = useHotMemoize(module, () => createEnv('kubernetes'));
   const devToolsEnv = useHotMemoize(module, () => createEnv('devtools'));
-  const onboardingEnv = useHotMemoize(module, () => createEnv('onboarding'));
+  const daskOnboardingEnv = useHotMemoize(module, () => createEnv('onboarding'));
 
   const apiRouter = Router();
   apiRouter.use('/catalog', await catalog(catalogEnv));
@@ -110,7 +110,7 @@ async function main() {
   apiRouter.use('/explore', await explore(exploreEnv));
   apiRouter.use('/kubernetes', await kubernetes(kubernetesEnv));
   apiRouter.use('/devtools', await devTools(devToolsEnv));
-  apiRouter.use('/dask-onboarding', await onboarding(onboardingEnv));
+  apiRouter.use('/dask-onboarding', await daskOnboarding(daskOnboardingEnv));
 
   // Add backends ABOVE this line; this 404 handler is the catch-all fallback
   apiRouter.use(notFoundHandler());
